BACKPORT: autofs-5.1.9 - don't free ext mount if mounted
If an external mount is in use when a umount is attempted don't free
it just let the reference count go to zero.

This will leave the mount in place and it won't get umounted. But if
another automount uses it it's reference count will become no zero
allowing for it to be umounted as normal if it isn't in use during
automount expire.

Signed-off-by: Ian Kent <raven@themaw.net>
Signed-off-by: ilikara <3435193369@qq.com>