BACKPORT: autofs-5.1.9 - fix deadlock in remount
If we're starting up or trying to re-connect to an existing direct mount
we could be iterating through the map entries with the cache readlock
held so we can't just take the writelock for direct mounts. But when
trying to re-connect to an existing mount at startup there won't be any
other process updating the map entry cache.

Signed-off-by: Ian Kent <raven@themaw.net>
Signed-off-by: ilikara <3435193369@qq.com>